+ Swagger, carService refactor, carController refactor
* Added swagger plugin to vehicle-api module
* Refactorized code in carService for better readability
* Refactorized code in carController for better readability

+++ vehicles-api/src/main/java/com/udacity/vehicles/service/CarService.java
@@ -25,10 +25,6 @@ public class CarService {
 	private final PriceClient priceClient;
 
 	public CarService(CarRepository repository, MapsClient mapsClient, PriceClient priceClient) {
-		/**
-		 * TODO: Add the Maps and Pricing Web Clients you create
-		 *   in `VehiclesApiApplication` as arguments and set them here.
-		 */
 		this.repository = repository;
 		this.mapsClient = mapsClient;
 		this.priceClient = priceClient;
@@ -58,28 +54,13 @@ public class CarService {
 
 		Car car = record.get();
 
-		/**
-		 * TODO: Use the Pricing Web client you create in `VehiclesApiApplication`
-		 *   to get the price based on the `id` input'
-		 * TODO: Set the price of the car
-		 * Note: The car class file uses @transient, meaning you will need to call
-		 *   the pricing service each time to get the price.
-		 */
 		String price = this.priceClient.getPrice(car.getId());
 
 		car.setPrice(price);
 
-		/**
-		 * TODO: Use the Maps Web client you create in `VehiclesApiApplication`
-		 *   to get the address for the vehicle. You should access the location
-		 *   from the car object and feed it to the Maps service.
-		 * TODO: Set the location of the vehicle, including the address information
-		 * Note: The Location class file also uses @transient for the address,
-		 * meaning the Maps service needs to be called each time for the address.
-		 */
-        Location fullLocation = this.mapsClient.getAddress(car.getLocation());
+		Location fullLocation = this.mapsClient.getAddress(car.getLocation());
 
-        car.setLocation(fullLocation);
+		car.setLocation(fullLocation);
 
 		return car;
 	}
@@ -109,7 +90,13 @@ public class CarService {
 	 * @param id the ID number of the car to delete
 	 */
 	public void delete(Long id) {
-        Car car = this.findById(id);
+		Optional<Car> record = this.repository.findById(id);
+
+		if (record.isEmpty()) {
+			throw new CarNotFoundException();
+		}
+
+		Car car = record.get();
 
 		this.repository.delete(car);
 	}
